Attempt to fix shilkas
ZSU-23-4 has the attribute `SAM TR`. The IADS script will classify a
group as a SAM as long as _any_ unit in the group has this attribute.

We don't want every random AAA or maneuver group with a shilka to be
classified as a SAM, so tweak the logic to only classify groups as a SAM
if at least one unit has the `SAM TR` attribute without also having the
`AAA` attribute.